Change object definitions to validate on proper integer classes

+++ flow_link.m
@@ -5,10 +5,10 @@ classdef flow_link
         diameter                    double      % pipe diameter, ft
         length                      double      % pipe length, ft
         delta_z                     double      % height differential end-over-end, ft
-        all_downstream_leaves(:, 1) int         % list of leaf indices
+        all_downstream_leaves(:, 1) int8        % list of leaf indices
         % Primary state parameters
-        pressure_iteration          int = 0;    % last time pressure values were recalculated
-        flow_iteration              int = 0;    % last time flow values were recalculated
+        pressure_iteration          uint8 = 0;  % last time pressure values were recalculated
+        flow_iteration              uint8 = 0;  % last time flow values were recalculated
         velocity                    double = 0; % fluid velocity, ft/s
         head_in                     double = 0; % heat feet at inlet, ft
         % Secondary state parameters
